Create Java Stdin & Stdout.java
In this code i have faced one problem that is very usual while coding.
Q. Why is Scanner skipping nextLine() after use of other next functions?
Ans. This is because the Scanner class of nextLine() method advances this scanner past the current line and returns the input that was skipped.This function prints the rest of the current line, leaving out the line separator at the end. The next is set to after the line separator. Since this method continues to search through the input looking for a line separator, it may search all of the input searching for the line to skip if no line separators are present.